style: remove reset-button in favor of search header
The "reset button" that was showing when no result wasn't allowing users to remove only one filter. The search header, which also contains a reset button, wasn't visible in case of no result. Yet this component allows more granular changes on the active filters. This also keep the reset button in the same place in every situations (with or without results).